src/directors_reimbursements/requirements.txt
appdirs==1.4.4
asttokens==3.0.0
clipboard==0.0.4
colorama==0.4.6
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.0
icecream==2.1.4
iniconfig==2.1.0
Markdown==3.7
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psi_toml==0.0.5
psiconfig==0.0.13
psiutils==0.1.49
Pygments==2.19.1
pyperclip==1.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
termcolor==2.5.0
tkinterhtml==0.7
tkinterweb==4.2.0
tomli==2.2.1
tomli_w==1.2.0
workbooky==0.0.5
